buildSrc: centralize per-platform build configuration for ffmpeg and mpv
Each module gets a single *Targets.kt as the sole source of its platform
variant configuration:

- FfmpegTargets.kt: common configure flags plus per-platform target factories
  (Windows x64/arm64, Linux, macOS, iOS, Android). FfmpegBuildTarget now also
  describes assemble-stage behavior (JNI wrapper name/link mode, MSYS2
  subsystem, DLL/TLS collection, install-name rewriting).
- MpvTargets.kt: common meson options plus per-platform factories.
  MpvBuildTarget gains MpvJniToolchain (compiler, args, link-library
  patterns, source extensions) and MpvRuntimeLayout (bin/lib layout plus a
  post-processing enum: collect DLLs / bundle dylibs / set RUNPATH / bundle
  libc++).

Task implementations become platform-agnostic data consumers: all
when(targetName)/startsWith(Macos) string dispatch in FfmpegAssembleTask,
MpvJniBuildTask and MpvAssembleTask is replaced by @Input properties fed from
the descriptors. Host dispatch collapses to one enabledTargets() function per
module; the identical apply-patch/snapshot/revert triplets are extracted to
nativebuild/PatchedSourceTemplate.kt. FfmpegSupport/MpvSupport shrink to
environment lookup contexts.

Also fixes Android cross builds on fresh Windows environments (verified by
building Windows x64 + Android arm64-v8a end to end; Windows runtime output
is file-for-file identical to before the refactor):
- FfmpegConfigureTask no longer runs pacman with an empty package list
  (Android ffmpeg has no MSYS2 packages; pacman errors on no targets).
- Meson >= 1.11 validates --prefix against the host system's path semantics
  on cross builds, rejecting Windows-style prefixes when targeting Android;
  cross builds now configure --prefix=/ and install via --destdir, with
  MSYS2_ARG_CONV_EXCL protecting the bare / from MSYS2 path conversion.
